+++ src/test/java/org/scion/jpan/api/DatagramChannelMultiSendPathTest.java
@@ -21,8 +21,8 @@ import java.nio.ByteBuffer;
 import java.nio.charset.Charset;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Test;
-import org.scion.jpan.DatagramChannel;
 import org.scion.jpan.Path;
+import org.scion.jpan.ScionDatagramChannel;
 import org.scion.jpan.ScionService;
 import org.scion.jpan.testutil.PingPongChannelHelper;
 
@@ -43,7 +43,7 @@ class DatagramChannelMultiSendPathTest {
     pph.runPingPong(serverFn, clientFn);
   }
 
-  private void client(DatagramChannel channel, Path serverAddress, int id) throws IOException {
+  private void client(ScionDatagramChannel channel, Path serverAddress, int id) throws IOException {
     String message = PingPongChannelHelper.MSG + "-" + id;
     ByteBuffer sendBuf = ByteBuffer.wrap(message.getBytes());
     channel.send(sendBuf, serverAddress);