+++ src/test/java/org/scion/jpan/api/DatagramChannelMultiWriteConnectedPathTest.java
@@ -21,9 +21,9 @@ import java.nio.ByteBuffer;
 import java.nio.charset.Charset;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Test;
-import org.scion.jpan.DatagramChannel;
 import org.scion.jpan.Path;
 import org.scion.jpan.RequestPath;
+import org.scion.jpan.ScionDatagramChannel;
 import org.scion.jpan.ScionService;
 import org.scion.jpan.testutil.PingPongChannelHelper;
 
@@ -46,7 +46,7 @@ class DatagramChannelMultiWriteConnectedPathTest {
     pph.runPingPong(serverFn, clientFn);
   }
 
-  private void client(DatagramChannel channel, RequestPath serverAddress, int id)
+  private void client(ScionDatagramChannel channel, RequestPath serverAddress, int id)
       throws IOException {
     String message = MSG + "-" + id;
     ByteBuffer sendBuf = ByteBuffer.wrap(message.getBytes());
@@ -65,7 +65,7 @@ class DatagramChannelMultiWriteConnectedPathTest {
     assertEquals(message, pong);
   }
 
-  private void server(DatagramChannel channel) throws IOException {
+  private void server(ScionDatagramChannel channel) throws IOException {
     ByteBuffer request = ByteBuffer.allocate(512);
     // System.out.println("SERVER: --- USER - Waiting for packet --------------------- " + i);
     Path path = channel.receive(request);