+++ src/test/java/org/scion/jpan/api/DatagramChannelStreamTest.java
@@ -22,8 +22,8 @@ import java.nio.charset.Charset;
 import java.util.ArrayList;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Test;
-import org.scion.jpan.DatagramChannel;
 import org.scion.jpan.Path;
+import org.scion.jpan.ScionDatagramChannel;
 import org.scion.jpan.ScionService;
 import org.scion.jpan.testutil.MockNetwork;
 import org.scion.jpan.testutil.PingPongChannelHelper;
@@ -48,7 +48,7 @@ class DatagramChannelStreamTest {
     assertEquals(2 * 2 * 2 * N_BULK, MockNetwork.getAndResetForwardCount());
   }
 
-  private void client(DatagramChannel channel, Path serverAddress, int id) throws IOException {
+  private void client(ScionDatagramChannel channel, Path serverAddress, int id) throws IOException {
     String message = PingPongChannelHelper.MSG + "-" + id;
     ByteBuffer sendBuf = ByteBuffer.wrap(message.getBytes());
 
@@ -80,7 +80,7 @@ class DatagramChannelStreamTest {
     }
   }
 
-  public void server(DatagramChannel channel) throws IOException {
+  public void server(ScionDatagramChannel channel) throws IOException {
     ByteBuffer request = ByteBuffer.allocate(512);
     //    System.out.println(
     //        "SERVER: --- USER - Waiting for packet --------------------- "