+++ src/test/java/org/scion/jpan/demo/ScmpEchoDemo.java
@@ -108,7 +108,7 @@ public class ScmpEchoDemo {
     ByteBuffer data = ByteBuffer.allocate(0);
 
     println("Listening on port " + localPort + " ...");
-    try (DatagramChannel channel = DatagramChannel.open()) {
+    try (ScionDatagramChannel channel = ScionDatagramChannel.open()) {
       channel.connect(path);
       println("Resolved local address: ");
       println("  " + channel.getLocalAddress().getAddress().getHostAddress());